Migrate victory-pie to TypeScript
