--- packages/victory-pie/src/index.js
@@ -1,2 +1,0 @@
-export { default as VictoryPie } from "./victory-pie";
-export { default as Slice } from "./slice";
