+++ packages/victory-pie/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./victory-pie";
+export * from "./slice";
